requirements.txt
# BuildVoice Dependencies
# =======================

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
jinja2==3.1.4

# Database
aiosqlite==0.20.0

# AI / NLP
groq>=0.12.0
rapidfuzz==3.9.7

# Speech-to-Text
vosk==0.3.45
sounddevice==0.5.5
pydub==0.25.1

# PDF Generation
reportlab==4.2.2

# Environment & Configuration
python-dotenv==1.0.1
PyJWT==2.11.0

# Utilities
pydantic==2.9.1
python-dateutil==2.9.0
loguru==0.7.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
